+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/electric/AbstractEnergyProvider.java
@@ -95,7 +95,7 @@ public abstract class AbstractEnergyProvider extends SlimefunItem implements Inv
     }
 
     @Nonnull
-    public Set<MachineFuel> getFuelTypes2() {
+    public Set<MachineFuel> getFuelTypes() {
         return fuelTypes;
     }